ref: properly split NweStudentForm from the main bundle size.
