+++ apps/vor/frontend/i18n/lingui/en/messages.po
@@ -32,7 +32,7 @@ msgstr "<0>\"{name}\"</0> curriculum and student data related to it will be perm
 
 #: src/components/LinkInput/LinkInput.tsx:9
 #: src/components/PageAddGuardian/PageAddGuardian.tsx:73
-#: src/components/PageNewStudent/NewStudentForm.tsx:122
+#: src/components/PageNewStudent/NewStudentForm.tsx:99
 #: src/components/PageNewStudentAddGuardian/PageNewStudentAddGuardian.tsx:70
 #: src/components/PageNewStudentPlans/PageNewStudentPlans.tsx:156
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:92
@@ -118,7 +118,7 @@ msgstr "Are you sure you want to delete this image?"
 msgid "Are you sure you want to delete {title}?"
 msgstr "Are you sure you want to delete {title}?"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:161
+#: src/components/PageNewStudent/NewStudentForm.tsx:138
 msgid "Are you sure you want to remove {0} from the list of guardians?"
 msgstr "Are you sure you want to remove {0} from the list of guardians?"
 
@@ -184,7 +184,7 @@ msgid "Class"
 msgstr "Class"
 
 #: src/components/PageClassSettings/PageClassSettings.tsx:27
-#: src/components/PageNewStudent/NewStudentForm.tsx:89
+#: src/components/PageNewStudent/NewStudentForm.tsx:66
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:66
 msgid "Classes"
 msgstr "Classes"
@@ -210,7 +210,7 @@ msgid "Create observation"
 msgstr "Create observation"
 
 #: src/components/EmptyClassDataPlaceholder/EmptyClassDataPlaceholder.tsx:7
-#: src/components/PageNewStudent/NewStudentForm.tsx:112
+#: src/components/PageNewStudent/NewStudentForm.tsx:89
 msgid "Create your first class to track your student's class enrollment."
 msgstr "Create your first class to track your student's class enrollment."
 
@@ -263,11 +263,11 @@ msgstr "Dark"
 msgid "Date"
 msgstr "Date"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:66
+#: src/components/PageNewStudent/NewStudentForm.tsx:43
 msgid "Date of Birth"
 msgstr "Date of Birth"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:69
+#: src/components/PageNewStudent/NewStudentForm.tsx:46
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:290
 msgid "Date of Entry"
 msgstr "Date of Entry"
@@ -434,7 +434,7 @@ msgstr "Export Observations"
 msgid "Father"
 msgstr "Father"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:77
+#: src/components/PageNewStudent/NewStudentForm.tsx:54
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:198
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:213
 msgid "Female"
@@ -457,7 +457,7 @@ msgstr "Friday"
 msgid "Gallery"
 msgstr "Gallery"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:72
+#: src/components/PageNewStudent/NewStudentForm.tsx:49
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:193
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:208
 msgid "Gender"
@@ -472,7 +472,7 @@ msgid "Give Feedback"
 msgstr "Give Feedback"
 
 #: src/components/EmptyClassDataPlaceholder/EmptyClassDataPlaceholder.tsx:7
-#: src/components/PageNewStudent/NewStudentForm.tsx:112
+#: src/components/PageNewStudent/NewStudentForm.tsx:89
 msgid "Go to Class Settings"
 msgstr "Go to Class Settings"
 
@@ -490,7 +490,7 @@ msgstr "Guardian Name"
 msgid "Guardian Profile"
 msgstr "Guardian Profile"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:118
+#: src/components/PageNewStudent/NewStudentForm.tsx:95
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:88
 msgid "Guardians"
 msgstr "Guardians"
@@ -554,7 +554,7 @@ msgstr "Log Out"
 msgid "Login for Teachers"
 msgstr "Login for Teachers"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:76
+#: src/components/PageNewStudent/NewStudentForm.tsx:53
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:196
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:212
 msgid "Male"
@@ -597,7 +597,7 @@ msgstr "Mother"
 msgid "Name"
 msgstr "Name"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:63
+#: src/components/PageNewStudent/NewStudentForm.tsx:40
 msgid "Name (Required)"
 msgstr "Name (Required)"
 
@@ -626,8 +626,8 @@ msgstr "New Guardian"
 msgid "New Observation"
 msgstr "New Observation"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:57
-#: src/components/PageNewStudent/PageNewStudent.tsx:27
+#: src/components/PageNewStudent/NewStudentForm.tsx:34
+#: src/components/PageNewStudent/PageNewStudent.tsx:28
 #: src/components/PageNewStudentAddGuardian/PageNewStudentAddGuardian.tsx:66
 #: src/components/StudentsList/StudentsList.tsx:80
 msgid "New Student"
@@ -701,7 +701,7 @@ msgid "None"
 msgstr "None"
 
 #: src/components/DateInput/DateInput.tsx:30
-#: src/components/PageNewStudent/NewStudentForm.tsx:75
+#: src/components/PageNewStudent/NewStudentForm.tsx:52
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:200
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:211
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:233
@@ -715,7 +715,7 @@ msgstr "Not Set"
 msgid "Note"
 msgstr "Note"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:82
+#: src/components/PageNewStudent/NewStudentForm.tsx:59
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:258
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:262
 msgid "Notes"
@@ -849,7 +849,7 @@ msgstr "Relationship"
 msgid "Remove Class?"
 msgstr "Remove Class?"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:161
+#: src/components/PageNewStudent/NewStudentForm.tsx:138
 msgid "Remove Guardian?"
 msgstr "Remove Guardian?"
 
@@ -869,7 +869,7 @@ msgstr "Repetition"
 #~ msgid "Reporting"
 #~ msgstr "Reporting"
 
-#: src/components/PageNewStudent/PageNewStudent.tsx:46
+#: src/components/PageNewStudent/PageNewStudent.tsx:47
 msgid "Reset Form"
 msgstr "Ulang Formulir"
 
@@ -888,7 +888,7 @@ msgstr "Saturday"
 #: src/components/PageNewClass/PageNewClass.tsx:80
 #: src/components/PageNewObservation/PageNewObservation.tsx:86
 #: src/components/PageNewSchool/PageNewSchool.tsx:50
-#: src/components/PageNewStudent/PageNewStudent.tsx:37
+#: src/components/PageNewStudent/PageNewStudent.tsx:38
 #: src/components/PageNewStudentClass/PageNewStudentClass.tsx:83
 #: src/components/PageNewSubject/PageNewSubject.tsx:117
 #: src/components/PageObservationDetails/DetailsCard.tsx:29
@@ -1015,7 +1015,7 @@ msgstr "Start with a blank curriculum that you can customize from scratch."
 msgid "Status"
 msgstr "Status"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:79
+#: src/components/PageNewStudent/NewStudentForm.tsx:56
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:233
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:237
 msgid "Student ID"
@@ -1025,7 +1025,7 @@ msgstr "Student ID"
 #: src/components/PageAddGuardian/PageAddGuardian.tsx:67
 #: src/components/PageAllObservations/PageAllObservations.tsx:13
 #: src/components/PageNewObservation/PageNewObservation.tsx:76
-#: src/components/PageNewStudent/PageNewStudent.tsx:26
+#: src/components/PageNewStudent/PageNewStudent.tsx:27
 #: src/components/PageNewStudentAddGuardian/PageNewStudentAddGuardian.tsx:65
 #: src/components/PageStudentImages/PageStudentImages.tsx:34
 #: src/components/PageStudentOverview/PageStudentOverview.tsx:24
@@ -1094,7 +1094,7 @@ msgstr "Theme"
 msgid "This guardian doesn't have any children registered yet."
 msgstr "This guardian doesn't have any children registered yet."
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:128
+#: src/components/PageNewStudent/NewStudentForm.tsx:105
 msgid "This student doesn't have a guardian yet."
 msgstr "This student doesn't have a guardian yet."
