+++ apps/vor/frontend/i18n/lingui/id/messages.po
@@ -37,7 +37,7 @@ msgstr "kurikulum <0>\"{name}\"</0> dan data murid terkait akan dihapus secara p
 
 #: src/components/LinkInput/LinkInput.tsx:9
 #: src/components/PageAddGuardian/PageAddGuardian.tsx:73
-#: src/components/PageNewStudent/NewStudentForm.tsx:122
+#: src/components/PageNewStudent/NewStudentForm.tsx:99
 #: src/components/PageNewStudentAddGuardian/PageNewStudentAddGuardian.tsx:70
 #: src/components/PageNewStudentPlans/PageNewStudentPlans.tsx:156
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:92
@@ -123,7 +123,7 @@ msgstr "Anda yakin ingin menghapus gambar ini?"
 msgid "Are you sure you want to delete {title}?"
 msgstr "Anda yakin mau menghapus {title}?"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:161
+#: src/components/PageNewStudent/NewStudentForm.tsx:138
 msgid "Are you sure you want to remove {0} from the list of guardians?"
 msgstr "Apa anda yakin ingin menghapus {0} dari daftar wali murid ini?"
 
@@ -189,7 +189,7 @@ msgid "Class"
 msgstr "Kelas"
 
 #: src/components/PageClassSettings/PageClassSettings.tsx:27
-#: src/components/PageNewStudent/NewStudentForm.tsx:89
+#: src/components/PageNewStudent/NewStudentForm.tsx:66
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:66
 msgid "Classes"
 msgstr "Kelas"
@@ -215,7 +215,7 @@ msgid "Create observation"
 msgstr "Catat observasi baru"
 
 #: src/components/EmptyClassDataPlaceholder/EmptyClassDataPlaceholder.tsx:7
-#: src/components/PageNewStudent/NewStudentForm.tsx:112
+#: src/components/PageNewStudent/NewStudentForm.tsx:89
 msgid "Create your first class to track your student's class enrollment."
 msgstr "Buat kelas pertama anda untuk melacak pendaftaran siswa."
 
@@ -268,11 +268,11 @@ msgstr "Gelap"
 msgid "Date"
 msgstr "Tanggal"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:66
+#: src/components/PageNewStudent/NewStudentForm.tsx:43
 msgid "Date of Birth"
 msgstr "Tanggal Lahir"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:69
+#: src/components/PageNewStudent/NewStudentForm.tsx:46
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:290
 msgid "Date of Entry"
 msgstr "Tanggal Masuk"
@@ -439,7 +439,7 @@ msgstr "Ekspor Observasi"
 msgid "Father"
 msgstr "Ayah"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:77
+#: src/components/PageNewStudent/NewStudentForm.tsx:54
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:198
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:213
 msgid "Female"
@@ -462,7 +462,7 @@ msgstr "Jumat"
 msgid "Gallery"
 msgstr "Galeri"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:72
+#: src/components/PageNewStudent/NewStudentForm.tsx:49
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:193
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:208
 msgid "Gender"
@@ -477,7 +477,7 @@ msgid "Give Feedback"
 msgstr "Beri kritik dan saran"
 
 #: src/components/EmptyClassDataPlaceholder/EmptyClassDataPlaceholder.tsx:7
-#: src/components/PageNewStudent/NewStudentForm.tsx:112
+#: src/components/PageNewStudent/NewStudentForm.tsx:89
 msgid "Go to Class Settings"
 msgstr "Ke Pengaturan Kelas"
 
@@ -495,7 +495,7 @@ msgstr "Nama Wali Murid"
 msgid "Guardian Profile"
 msgstr "Profil Wali"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:118
+#: src/components/PageNewStudent/NewStudentForm.tsx:95
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:88
 msgid "Guardians"
 msgstr "Wali Murid"
@@ -559,7 +559,7 @@ msgstr "Keluar"
 msgid "Login for Teachers"
 msgstr "Login untuk guru"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:76
+#: src/components/PageNewStudent/NewStudentForm.tsx:53
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:196
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:212
 msgid "Male"
@@ -602,7 +602,7 @@ msgstr "Ibu"
 msgid "Name"
 msgstr "Nama"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:63
+#: src/components/PageNewStudent/NewStudentForm.tsx:40
 msgid "Name (Required)"
 msgstr "Nama (Wajib)"
 
@@ -631,8 +631,8 @@ msgstr "Wali Baru"
 msgid "New Observation"
 msgstr "Observasi Baru"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:57
-#: src/components/PageNewStudent/PageNewStudent.tsx:27
+#: src/components/PageNewStudent/NewStudentForm.tsx:34
+#: src/components/PageNewStudent/PageNewStudent.tsx:28
 #: src/components/PageNewStudentAddGuardian/PageNewStudentAddGuardian.tsx:66
 #: src/components/StudentsList/StudentsList.tsx:80
 msgid "New Student"
@@ -706,7 +706,7 @@ msgid "None"
 msgstr "Tidak ada"
 
 #: src/components/DateInput/DateInput.tsx:30
-#: src/components/PageNewStudent/NewStudentForm.tsx:75
+#: src/components/PageNewStudent/NewStudentForm.tsx:52
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:200
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:211
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:233
@@ -720,7 +720,7 @@ msgstr "Belum Dipilih"
 msgid "Note"
 msgstr "Catatan"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:82
+#: src/components/PageNewStudent/NewStudentForm.tsx:59
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:258
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:262
 msgid "Notes"
@@ -854,7 +854,7 @@ msgstr "Hubungan"
 msgid "Remove Class?"
 msgstr "Hapus Kelas?"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:161
+#: src/components/PageNewStudent/NewStudentForm.tsx:138
 msgid "Remove Guardian?"
 msgstr "Hilangkan Wali Murid?"
 
@@ -874,7 +874,7 @@ msgstr "Pengulangan"
 #~ msgid "Reporting"
 #~ msgstr "Laporan"
 
-#: src/components/PageNewStudent/PageNewStudent.tsx:46
+#: src/components/PageNewStudent/PageNewStudent.tsx:47
 msgid "Reset Form"
 msgstr "Ulang Formulir"
 
@@ -893,7 +893,7 @@ msgstr "Sabtu"
 #: src/components/PageNewClass/PageNewClass.tsx:80
 #: src/components/PageNewObservation/PageNewObservation.tsx:86
 #: src/components/PageNewSchool/PageNewSchool.tsx:50
-#: src/components/PageNewStudent/PageNewStudent.tsx:37
+#: src/components/PageNewStudent/PageNewStudent.tsx:38
 #: src/components/PageNewStudentClass/PageNewStudentClass.tsx:83
 #: src/components/PageNewSubject/PageNewSubject.tsx:117
 #: src/components/PageObservationDetails/DetailsCard.tsx:29
@@ -1020,7 +1020,7 @@ msgstr "Mulai dengan kurikulum kosong yang dapat Anda kustom dari awal."
 msgid "Status"
 msgstr "Status"
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:79
+#: src/components/PageNewStudent/NewStudentForm.tsx:56
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:233
 #: src/components/PageStudentProfile/PageStudentProfile.tsx:237
 msgid "Student ID"
@@ -1030,7 +1030,7 @@ msgstr "ID Murid"
 #: src/components/PageAddGuardian/PageAddGuardian.tsx:67
 #: src/components/PageAllObservations/PageAllObservations.tsx:13
 #: src/components/PageNewObservation/PageNewObservation.tsx:76
-#: src/components/PageNewStudent/PageNewStudent.tsx:26
+#: src/components/PageNewStudent/PageNewStudent.tsx:27
 #: src/components/PageNewStudentAddGuardian/PageNewStudentAddGuardian.tsx:65
 #: src/components/PageStudentImages/PageStudentImages.tsx:34
 #: src/components/PageStudentOverview/PageStudentOverview.tsx:24
@@ -1099,7 +1099,7 @@ msgstr "Tema"
 msgid "This guardian doesn't have any children registered yet."
 msgstr "Wali murid ini belum memiliki data murid."
 
-#: src/components/PageNewStudent/NewStudentForm.tsx:128
+#: src/components/PageNewStudent/NewStudentForm.tsx:105
 msgid "This student doesn't have a guardian yet."
 msgstr "Murid ini belum mempunyai wali."
