+++ apps/vor/frontend/src/components/Layout/Layout.tsx
@@ -9,11 +9,11 @@ import {
 } from "../../hooks/schoolIdState"
 import Navbar from "../Navbar/Navbar"
 import { useGetUserProfile } from "../../hooks/api/useGetUserProfile"
+import { NewStudentFormProvider } from "../PageNewStudent/NewStudentFormContext"
 import Typography from "../Typography/Typography"
 import StudentsList from "../StudentsList/StudentsList"
 import { borderRight } from "../../border"
 import { navigate } from "../Link/Link"
-import { NewStudentFormProvider } from "../PageNewStudent/NewStudentForm"
 
 /** Top level component which encapsulate most pages. Provides Appbar and Sidebar for navigation.
  *
