+++ apps/vor/frontend/src/components/PagePickGuardian/PagePickGuardian.tsx
@@ -6,6 +6,7 @@ import { navigate } from "../Link/Link"
 import BackNavigation from "../BackNavigation/BackNavigation"
 import { NEW_STUDENT_URL } from "../../routes"
 import Input from "../Input/Input"
+import { useNewStudentFormContext } from "../PageNewStudent/NewStudentFormContext"
 import Select from "../Select/Select"
 import { GuardianRelationship } from "../../hooks/api/students/usePostNewStudent"
 import Typography from "../Typography/Typography"
@@ -21,7 +22,6 @@ import {
 } from "../../hooks/api/guardians/useGetSchoolGuardians"
 import TextArea from "../TextArea/TextArea"
 import GuardianRelationshipPickerDialog from "../GuardianRelationshipPickerDialog/GuardianRelationshipPickerDialog"
-import { useNewStudentFormContext } from "../PageNewStudent/NewStudentForm"
 
 export const PagePickGuardian: FC = () => {
   const { i18n } = useLingui()
